Migrate `PluginInstall` plugin logics into standalone command
