+++ lib/plugins/index.js
@@ -16,7 +16,6 @@ module.exports = [
   require('./rollback.js'),
   require('./slstats.js'),
   require('./plugin/plugin.js'),
-  require('./plugin/install.js'),
   require('./plugin/uninstall.js'),
   require('./plugin/list.js'),
   require('./plugin/search.js'),
